Working ANSI fore color parsing as Text / TextColored
