+++ doc/examples/integrations_pyglet.py
@@ -39,15 +39,11 @@ def main():
         imgui.begin("Custom window", True)
         imgui.text("Bar")
         imgui.text_colored("Eggs", 0.2, 1., 0.)
-        # colorstr = colorama.an  \033[31m
-        # imgui.text_ansi("BarAnsi: "+Fore.RED+" whoop")
-        # # print("BarAnsi: "+Fore.RED+" whoop")
-        # imgui.text_ansi_colored("EggsAnsiColored: "+Fore.RED+" whoop", 0.2, 1., 0.)
-        message = "BarAnsi: "+"\033[31m" +" whoop"+"\033[0m"
-        print(message)
-        imgui.text_ansi("BarAnsi: "+"\033[31m" +" whoop"+"\033[0m")
+        
+        # imgui.text_ansi("BarAnsi: "+"\033[31m" +" whoop"+colorama.Fore.RESET+"\033[m")
+        imgui.text_ansi("B\033[31marA\033[mnsi ")
         # print("BarAnsi: "+"\033[31m"+" whoop"+"\033[0m")
-        imgui.text_ansi_colored("EggsAnsiColored: "+'\033[31m' +" whoop"+"\033[0m", 0.2, 1., 0.)
+        imgui.text_ansi_colored("Eg\033[31mgAn\033[msi ", 0.2, 1., 0.)
 
         imgui.end()
